Add a new "accept" endpoint prior to signing up.
This change introduces a new endpoint (/web/accept/:tokenId) that simply
prompts the user to create an account by clicking a button, when they do
so they will be redirected to /web/invite/:tokenId (which is what is
used currently)

This change is aimed at preventing web crawlers/proxies and malware link
scanners in email clients that are hitting the /invite/ endpoint
directly shortly after the user clicks the link, which invalidates the
URL.

Please see https://github.com/gravitational/cloud/issues/1079 for
further investigation and justification.

It is proposed that this link will be used for any outbound email
communication, or other situations where it's plausible that clients may
interfere with the invite link. It will not replace the invite link
entirely.