Update protobufs for Teleterm (LoginRequest params)
This commit removes a bunch of unused protobufs and also updates some of
those that got updated on the teleterm branch in the teleport repo.

The biggest change from all of them is that LoginRequest now has oneof
on Sso and Local. [1] This is because a login request should have either
Sso or Local params, but not both at the same time.

The previous implementation called both `setSso` and `setLogin` on the same
object. This no longer works with the use of `oneof`, because calling `setLogin`
after `setSso` would clear the `Sso` params. [2]

[1] https://github.com/gravitational/teleport/pull/10286#discussion_r821743202
[2] https://developers.google.com/protocol-buffers/docs/proto3#oneof_features