Fix getting cwd in presence of lsof warnings
lsof uses stderr to print out warnings. This caused the previous version
of ptyProcess to throw an error, even though the exit code of lsof was 0.

The existing code already handles non-zero exit codes (asyncExec will
reject the promise). Since we don't know why stderr was inspected in
the first place, let's just remove the check for it.