[v9] Backport Teleport Connect (3 of 5) (#764)
* Limit navigation capabilities to reduce attack surface

At the moment we don't create new windows nor navigate away from the rendered
app, so we can just block everything.

* Update to electron@13.6.9 (#703)

* Use x64 arch when building & packaging Teleterm

Our build system doesn't support arm64 for Mac releases yet (see issue
gravitational/teleport#4226 for more information). Because of that, for
the preview release we're likely going to have only the x64 version of
Teleterm.

This means that the shipped version of tsh should also be the x64 version.

I tried to change electron-builder's config to use x64 for macOS, but the
config options don't seem to work. I tried `mac.defaultArch` as well as
changing `mac.target` in various ways but `electron-builder install-app-deps`
just doesn't pick up those options. Both were set through
`packages/teleterm/package.json`

* Add `Notifications` component and service

* Show errors in `ClusterResources`' tables using standard `Danger` labels

* Use `Notifications` error in `syncRootCluster()` and `removeGateway()`

* Do not block app rendering when initializing function fails

* Fix accessing `serversSyncStatus` Map in `clustersService`

* Revert "Use x64 arch when building & packaging Teleterm"

This reverts commit 276e9a97402bdffce2e0fbea2a781d51c30cbce3.

Turns out that for development, we need to use arm64 version of native
deps. The build server is going to use x64 anyway, as per the reverted
commit, but when making manual demo builds, we'll have to remember to
use x64 for Teleterm and tsh.

* Submit modals' forms on `Enter` press

* Remove global `keyDown` handler from `KeyboardArrowsNavigation` as it blocked submitting forms

* Use teleterm/logger in runtimeSettings (#716)

The one from shared/libs/logger calls `window`, which doesn't exist in
the context of Electron main process.

* Improve Teleterm README (#719)

* Mention that `yarn build-term` needs to be run first before attempting
  to run the app in dev mode (already had a couple of people who had
  problems with setting up the app because they didn't run this first).
* Mention the assumption about both repos living in the same folder.
* Move the architecture diagram to the end of the file. Most people
  reading the README are not doing it for the diagram, but build
  instructions.
* Explain when gRPC files need to be recompiled.

* Prevent crash when network or cluster is offline (#712)

* Simplify the db connection tab (#720)

As described in gravitational/webapps.e#177, we want to replace the db
tab with just two sections:

* "Connect with CLI" which will show the command to use in terminal
* "Connect with GUI" which links to our documentation

After gravitational/teleport#11720 gets merged, the "Connect with CLI"
section will be massively simplified: it'll be basically just something like:

    psql postgres://localhost:12345

Moreover, @smallinsky suggested that tsh should be responsible for creating
those CLI connection commands. We should also do this in the future as
it'll let us support new protocols as soon as they land in tsh, without
us having to touch Teleterm codebase, for the most part.

* Resolve shell env (#718)

Co-authored-by: Grzegorz Zdunek <gzdunek@users.noreply.github.com>
Co-authored-by: Grzegorz Zdunek <grzegorz.zdunek@goteleport.com>