Change page size in Connect to 15
Just like in the Web UI, Connect is able to easily accommodate 15 resources
on the screen without vertical scroll in both 1920x1080 and 3024x1964,
as long as labels for the resources span just a single line.

1920x1080 remains the most popular screen size on desktop [1] and 1440x900
seems to be the most popular size on macOS [2].

On 1440x900 the pagination controls are not shown on the screen.

[1] https://gs.statcounter.com/screen-resolution-stats/desktop/north-america/#monthly-202105-202205
[2] https://web.archive.org/web/20220630103803/https://store.steampowered.com/hwsurvey/Steam-Hardware-Software-Survey-Welcome-to-Steam?platform=mac