Add automation to update webassets on push (#868)
This will run using a service account which is the only account
permitted to push to webassets and webassets.e.

The action will run when any commits are pushed to the protected
branches in webapps. This means that commits to webapps.e do not
directly trigger updates - this automation will run when the
enterprise submodule ref is updated in webapps.

The "nop" yarn script allows us to separate the building of the docker
image from the build of Teleport, so that GitHub actions can run the
yarn build after buildling an image from our dockerfile. This is
necessary because container steps do not support build args.