Remove custom entitlements
electron-builder adds some default ones if not specified and they're just
enough for us. On top of that, by letting electron-builder handle that
we make sure that they'll stay up to date in case they ever need to be
updated due to changes in Electron or Chromium. See [1] for an example.

[1] https://www.electronjs.org/blog/electron-6-0#electron-helper-rendererapp-electron-helper-gpuapp-and-electron-helper-pluginapp