Allow for resources to specify `shouldPrompt` and an optional `wrapper`.
Wrappers allow us to wrap the entire resource flow in any context
providers that it needs, without wrapping the entire Discover flow
itself in those contexts.

`shouldPrompt` allows the resource to decide based on the current step
number, if it should prompt the user when they go to navigate away.