retryWithRelogin: Enable use outside of document context
So far retryWithRelogin has assumed that it's used within a document.
This was manifested by the fact that in order to check if the relevant UI
is still active when the response comes in, retryWithRelogin accepted
originatingDocumentUri to check if that document is still visible on the
screen.

In access request PRs, we want to use retryWithRelogin in places outside
of the context of a document. The question of whether the relevant UI
is still active won't be answered by checking the status of a document
but rather through other means.

This commit makes it possible to express that in the code. Instead of taking
a document URI as an argument, retryWithRelogin now takes a function which
answers said question.

A helper function was added for convenience for places which still use
retryWithRelogin within the context of a document.