Now that clipboard.readText() requires transient user activation (see https://developer.mozilla.org/en-US/docs/Web/API/Clipboard/readText#security), our system of syncing the local clipboard to remote onfocus and onmouseenter no longer works, since those events don't necessarily imply a transient event.
With this change, we instead sendLocalClipboardToRemote onMouseDown and onKeyDown,
which should guarantee that the sync is always within a transient event.
Because the user must preform at least one of these actions in order to
initiate a paste, the clipboard should appear to always be in sync from
their perspective.