eliminates the entire useClipboard hook and its surrounding state, which was a big rube goldberg machine ensuring that 1. If a user had clipboard sharing enabled by RBAC 2. they were on chrome and 3. they had clipboard permissions allowed
If 1 was true, but either of 2 or 3 wasn’t, we blocked the UI and forced
them to correct it. That was also the mechanism that caused the clipboard
permissions dialog to pop up at the beginning of the session if clipboard
access wasn’t "granted".

Now the UX is much simpler: if a user has clipboard sharing allowed by RBAC,
and they’re on Chrome, clipboard sharing is enabled. If either of those is false,
clipboard sharing is disabled. If a user’s clipboard permissions are in
"prompt" state (aka neither "granted" nor "denied"), the first time they
try to copy/cut something on the remote, or have something in their clipboard
locally and click a mouse button or press a key, the browser prompt asking
for clipboard permissions will show up.