Connect: Adjust size of sync button & search input (#1280)
The size of the sync button was changed to match the refresh button of access requests (https://github.com/gravitational/webapps.e/pull/396#discussion_r982842537). The change was reverted in #1269.

Instead of making the sync button to be of custom height, I think we should just change the size of the search input.

It also has the added bonus of making the sync and refresh button be on a matching height on both tabs, so they don't shift when switching between access requests and cluster resources.