Make linguist correctly classify JS protobuf files
Linguist seems to expect that on line 6, those files have a specific string
in them but that's not true for our files.

Maybe that's because we use buf and not protoc to generate them.

This should automatically collapse those files in diffs and properly count
JavaScript usage in language stats for the repo.

See: https://github.com/github/linguist/blob/master/docs/overrides.md