Update ensureBaseUrl conditional
Previously we would only check if the `cfg.baseUrl` was the 0th index of the url being passed in during redirects.
This would catch most 'bad' redirects except for ones that would use the entire baseUrl as subdomains.
For example,
if `cfg.baseUrl` was `https://teleport.avat.us`, I could perform an open redirect by passing a
redirect_uri of `https://teleport.avat.us.bad.io` because the baseUrl was still the 0th index of the passed url.
This fix takes advantage of the URL api by passing in `cfg.baseUrl` as the optional base param in URL.

If a full, valid url is passed as the first param into `new URL`, any base param passed as the second will be overridden.
If thats the case, we can check if it is still equal to our `cfg.baseUrl` (a normal redirect will be), and prevent
any sneaky redirects from bad actors.