Fix websocket close
In #1442 we added a close code when closing websockets, but didn't
properly adhere to the API. When explicitly setting a close code,
you must either use the normal code or use a custom code. Usage of
the standard codes in the 1001-1015 range is reserved for the browser.

Fixes gravitational/teleport#19688